Linux/GCC related fixes (#1)
* linux fixes

Fixes case sensitive path, -1 used in a uint16_t enum, CMD_F and soh include path

* fix type correctness

+++ OTRExporter/RoomExporter.cpp
@@ -172,7 +172,7 @@ void OTRExporter_Room::Save(ZResource* res, const fs::path& outPath, BinaryWrite
 			for (const auto& c : cmdCsCam->cameras) {
 				writer->Write(c.type);
 				writer->Write(c.numPoints);
-				for (size_t i = 0; i < c.numPoints; i++) {
+				for (int16_t i = 0; i < c.numPoints; i++) {
 					writer->Write(cmdCsCam->points[((c.baseOffset - arrBase) / 6) + i].scalars[0].scalarData.s16);
 					writer->Write(cmdCsCam->points[((c.baseOffset - arrBase) / 6) + i].scalars[1].scalarData.s16);
 					writer->Write(cmdCsCam->points[((c.baseOffset - arrBase) / 6) + i].scalars[2].scalarData.s16);
